requirements.txt
beautifulsoup4
pandas
requests
github3.py
config-parser
selenium
